Web:Project Page:Delete Search:Trigger reload sections instead of page
On Search Delete, the various sections that need to update do a reload
from server instead of reloading the whole page.

+++ limelight_webapp/front_end/src/js/page_js/data_pages/other_data_pages/project_page/project_page_main_page_react_based/jsx/projectPage_SearchEntry_UsedInMultipleSections_Component.tsx
@@ -284,7 +284,14 @@ export class ProjectPage_SearchEntry_UsedInMultipleSections_Component extends Re
 
         const deleteComplete_Callback = (): void => {
 
-            this.props.deleteSearch_Callback({ projectSearchId: this.props.searchDisplayListItem.projectSearchId });
+            if ( this.props.deleteSearch_Callback ) {
+
+                this.props.deleteSearch_Callback( { projectSearchId: this.props.searchDisplayListItem.projectSearchId } );
+
+                return; // EARLY RETURN
+            }
+
+            window.location.reload(true)
         }
 
         this.props.projectPage_SearchesAdmin.deleteSearch({
